+++ src/decorators/input-type/zod-args.ts
@@ -9,6 +9,54 @@ import type { AnyZodObject, infer as Infer } from 'zod'
 
 type PT<F = any, T = any> = PipeTransform<F, T> | Type<PipeTransform<F, T>>
 
+let GENERATED_TYPES: WeakMap<AnyZodObject, object> | undefined
+let USED_NAMES: string[] | undefined
+
+/**
+ * Creates a new type from given zod object or returns previously created one.
+ *
+ * @template T The type of the zod object passed.
+ * @param {T} input The zod scheme object.
+ * @return {*} The newly or previously created class instance.
+ */
+function _getOrCreateRegisteredType<T extends AnyZodObject>(input: T) {
+  if (!GENERATED_TYPES) { GENERATED_TYPES = new WeakMap() }
+  let RegisteredType = GENERATED_TYPES.get(input) as Type<Infer<T>> | undefined
+  if (RegisteredType) return RegisteredType
+
+  const { name, description } = extractNameAndDescription(input, {})
+  const safeName = _getSafeName(name)
+  RegisteredType = inputFromZod(input, { name: safeName, description })
+  GENERATED_TYPES.set(input, RegisteredType)
+  return RegisteredType
+}
+
+/**
+ * Checks if the name is used before, in that case, adds a suffix of `_{number}`
+ * indicating the number of times the name is used.
+ *
+ * @param {string} name The name to check.
+ * @return {string} The name that is not used in any other types before.
+ */
+function _getSafeName(name: string): string {
+  if (!USED_NAMES) { USED_NAMES = [] }
+
+  let total = 0
+  for (let i = 0, limit = USED_NAMES.length; i < limit; ++i) {
+    const current = USED_NAMES[ i ]
+    if (current.startsWith(name)) { ++total }
+  }
+
+  if (total) {
+    const newName = `${name}_${total + 1}`
+    USED_NAMES.push(newName)
+    return newName
+  }
+
+  USED_NAMES.push(name)
+  return name
+}
+
 /**
  * A parameter decorator that takes a `zod` validation input and marks it as
  * GraphQL `Args` with `property` name with given `options` and pipes.
@@ -18,11 +66,11 @@ type PT<F = any, T = any> = PipeTransform<F, T> | Type<PipeTransform<F, T>>
  * @param {T} input The `zod` validation schema object.
  * @param {string} property The name of the property for the GraphQL request
  * argument.
- * 
+ *
  * @param {ArgsOptions} options The options for {@link Args} decorator.
  * @param {...PT[]} pipes The pipes that will be passed to {@link Args}
  * decorator.
- * 
+ *
  * @return {ParameterDecorator} A {@link ParameterDecorator} for GraphQL
  * argument.
  */
@@ -43,7 +91,7 @@ export function ZodArgs<T extends AnyZodObject>(
  * @param {ArgsOptions} options The options for {@link Args} decorator.
  * @param {...PT[]} pipes The pipes that will be passed to {@link Args}
  * decorator.
- * 
+ *
  * @return {ParameterDecorator} A {@link ParameterDecorator} for GraphQL
  * argument.
  */
@@ -62,10 +110,10 @@ export function ZodArgs<T extends AnyZodObject>(
  * @param {T} input The `zod` validation schema object.
  * @param {string} property The name of the property for the GraphQL request
  * argument.
- * 
+ *
  * @param {...PT[]} pipes The pipes that will be passed to {@link Args}
  * decorator.
- * 
+ *
  * @return {ParameterDecorator} A {@link ParameterDecorator} for GraphQL
  * argument.
  */
@@ -84,7 +132,7 @@ export function ZodArgs<T extends AnyZodObject>(
  * @param {T} input The `zod` validation schema object.
  * @param {...PT[]} pipes The pipes that will be passed to {@link Args}
  * decorator.
- * 
+ *
  * @return {ParameterDecorator} A {@link ParameterDecorator} for GraphQL
  * argument.
  */
@@ -131,10 +179,7 @@ export function ZodArgs<T extends AnyZodObject>(
     }
   }
 
-  // Operation
-  const { name, description } = extractNameAndDescription(input, {})
-  const RegisteredType = inputFromZod(input, { name, description })
-
+  const RegisteredType = _getOrCreateRegisteredType(input)
   pipes.unshift(new ZodValidatorPipe(input, RegisteredType))
 
   options ??= {}
@@ -165,4 +210,20 @@ export module ZodArgs {
    * A type for inferring the type of a given `zod` validation object.
    */
   export type Of<T extends AnyZodObject> = Infer<T>
+
+  /**
+   * Frees the used objects during the startup.
+   *
+   * The {@link ZodArgs} decorator uses helper local variables to keep the
+   * naming system working when the same scheme is used multiple times in
+   * separate decorators and same name for different schemes.
+   *
+   * This function should be called after the GraphQL scheme is created.
+   *
+   * @export
+   */
+  export function free() {
+    USED_NAMES = undefined
+    GENERATED_TYPES = undefined
+  }
 }